move logic to common viewmodel
